+++ jyra/bot/middleware/__init__.py
@@ -1,0 +1,9 @@
+"""Middleware package for Jyra bot.
+
+This package contains middleware functions that can be applied to bot handlers.
+"""
+
+from jyra.bot.middleware.context_middleware import context_middleware
+from jyra.bot.middleware.rate_limit_middleware import rate_limit_middleware
+
+__all__ = ['context_middleware', 'rate_limit_middleware']